+++ .examples/backend/src/endpoint/apis/contents/contents.module.ts
@@ -1,4 +1,4 @@
-import { AdapterModule, Module, ConfigModule } from "@zum-portal-core/backend";
+import { AdapterModule, Module, ConfigModule } from "@zum-front-core/backend";
 import { ContentsService } from "./contents.service";
 import { ContentsController } from "./contents.controller";