+++ packages/backend/src/base/adapter/zum.provision.adapter.ts
@@ -74,8 +74,9 @@ export class ZumProvisionAdapter {
       }) as AxiosPromise<T>;
     }
 
-    return Axios(
+    return Axios.request(
       deepmerge(option, {
+        method,
         timeout: option.timeout || 60000,
         headers: version ? { Accept: `application/vnd.zum.resource-${version}+json` } : {},
       }),