+++ packages/frontend/.eslintrc.js
@@ -3,5 +3,5 @@ module.exports = {
     node: true,
     browser: true,
   },
-  extends: ['@zum-portal-core/eslint-config-zum/vue', 'plugin:prettier/recommended'],
+  extends: ['@zum-front-core/eslint-config-zum/vue', 'plugin:prettier/recommended'],
 };