+++ packages/backend/CHANGELOG.md
@@ -1,7 +1,11 @@
 # CHANGELOG
 
+## 1.1.6
+- fix: Axios.request에 method 추가 [45f26d6](https://git.zuminternet.com/zum-portal-framework/zum-portal-core-js-project/commit/45f26d60f4974b91cb3f7fce7ee87c427c793013) junil
+- fix: changelog 수정 [1f531cf](https://git.zuminternet.com/zum-portal-framework/zum-portal-core-js-project/commit/1f531cf0ff2ecbc15a678338bf90dd3fa8d70f8d) junil
+
 ## 1.1.5
-- fix: adapter에서 post, put 등을 사용할 수 있도록 수정 [ca837cd](https://git.zuminternet.com/zum-portal-framework/zum-portal-core-js-project/commit/ca837cd320c27eca2c01d1b3a751c2b99fb1632f) junil
+- fix: adapter에서 axios.request 사용하도록 작업 완료 [23f4378](https://git.zuminternet.com/zum-portal-framework/zum-portal-core-js-project/commit/23f437872a8f67c91c41849492c69e2465719ebf) junil
 
 ## 1.1.4
 - feat: nest application options 를 주입할 수 있도록 작업 완료 [a1882af](https://git.zuminternet.com/zum-portal-framework/zum-portal-core-js-project/commit/a1882afa54135437393cf6a88e5bff34a8d5dad1) junil